+++ packages/vike-server/src/config.ts
@@ -9,16 +9,38 @@ export { config as default }
 const config = {
   name: 'vike-server',
   require: {
-    vike: '>=0.4.228'
+    vike: '>=0.4.231',
+    ['vike-react']: {
+      version: '>=0.6.4',
+      optional: true
+    },
+    ['vike-vue']: {
+      version: '>=0.9.2',
+      optional: true
+    },
+    ['vike-solid']: {
+      version: '>=0.7.11',
+      optional: true
+    }
   },
   vite: {
     plugins: [photon(), vikeServer()]
   },
   vite6BuilderApp: true,
+  // @ts-ignore
+  stream: {
+    enable: null,
+    type: 'web'
+  },
   meta: {
     server: {
       env: { config: true },
       global: true
+    },
+    // +stream is defined by vike-{react,vue,solid} but we define it again here to avoid Vike throwing the "unknown config" error if the user doesn't use vike-{react,vue,solid}
+    stream: {
+      env: { config: true },
+      isDefinedByPeerDependency: true
     }
   }
 } satisfies Config