+++ packages/vike-server/src/plugin/index.ts
@@ -1,14 +1,17 @@
+export { vikeServer, vikeServer as default }
+
 import { installPhoton } from '@photonjs/core/vite'
 import type { Plugin } from 'vite'
+import { configPlugin } from './plugins/configPlugin.js'
 import { serverEntryPlugin } from './plugins/serverEntryPlugin.js'
 import { setPhotonMeta } from './plugins/setPhotonMeta.js'
 import { standalonePlugin } from './plugins/standalonePlugin.js'
 import { vikeServerConfigToPhotonPlugin } from './plugins/vikeServerConfigToPhotonPlugin.js'
 
-export { vikeServer, vikeServer as default }
-
-function vikeServer(): Plugin[] {
+type PluginInterop = Record<string, unknown> & { name: string }
+function vikeServer(): PluginInterop[] {
   return [
+    configPlugin(),
     vikeServerConfigToPhotonPlugin(),
     ...serverEntryPlugin(),
     standalonePlugin(),
@@ -18,5 +21,6 @@ function vikeServer(): Plugin[] {
         return 'vike-server/universal-middlewares'
       }
     })
-  ]
+    // biome-ignore lint/suspicious/noExplicitAny: cast
+  ] satisfies Plugin[] as any
 }