+++ packages/vike-server/src/plugin/plugins/configPlugin.ts
@@ -1,0 +1,14 @@
+import type { Plugin } from 'vite'
+
+export function configPlugin(): Plugin {
+  return {
+    name: 'vike-server:common-config',
+    config() {
+      return {
+        resolve: {
+          noExternal: ['vike-server']
+        }
+      }
+    }
+  }
+}